+++ src/test/converter2/issues/gh2917.ts
@@ -1,0 +1,10 @@
+export interface Foo {
+    data: {
+        [key: string]: any;
+    };
+    mixed: {
+        (): string;
+        a: string;
+        [key: string]: any;
+    };
+}